+++ frontend/wailsjs/go/app/App.d.ts
@@ -12,6 +12,8 @@ export function Aggregate(arg1:string,arg2:string,arg3:string,arg4:string,arg5:s
 
 export function AskConfirmation(arg1:string):Promise<boolean>;
 
+export function CountItems(arg1:string,arg2:string,arg3:string,arg4:string):Promise<app.CountItemsResult>;
+
 export function CreateIndex(arg1:string,arg2:string,arg3:string,arg4:string):Promise<string>;
 
 export function DropCollection(arg1:string,arg2:string,arg3:string):Promise<boolean>;